Restructure to PC folder, functions in CitationAnalysis
