+++ tests/test_integration.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 import pytest
-import PapersCited
+import PapersCited.citationAnalysis as ca
 import os
 
 import locale
@@ -8,30 +8,30 @@ locale.setlocale(locale.LC_ALL, "")
 
 @pytest.mark.xfail(reason = "VSCode Pytest issue with ANSI encoding. In terminal, this should xpass")
 def test_read_analyze_text(rootdir):
-    text = PapersCited.read_document(os.path.join(rootdir, "sample_text.txt"))
+    text = ca.read_document(os.path.join(rootdir, "sample_text.txt"))
     
     # Get all types of citations
-    solo_authors = PapersCited.get_matches_solo_author(text, drop_excluded_phrases = True)
-    two_authors = PapersCited.get_matches_two_authors(text, drop_excluded_phrases = True)
-    three_authors = PapersCited.get_matches_three_authors(text, drop_excluded_phrases = True)
-    author_et_al = PapersCited.get_matches_author_et_al(text, drop_excluded_phrases = True)
-    two_surnames = PapersCited.get_matches_two_surnames(text, drop_excluded_phrases = True)
-    two_surnames_et_al = PapersCited.get_matches_two_surnames_et_al(text, drop_excluded_phrases = True)
+    solo_authors = ca.get_matches_solo_author(text, drop_excluded_phrases = True)
+    two_authors = ca.get_matches_two_authors(text, drop_excluded_phrases = True)
+    three_authors = ca.get_matches_three_authors(text, drop_excluded_phrases = True)
+    author_et_al = ca.get_matches_author_et_al(text, drop_excluded_phrases = True)
+    two_surnames = ca.get_matches_two_surnames(text, drop_excluded_phrases = True)
+    two_surnames_et_al = ca.get_matches_two_surnames_et_al(text, drop_excluded_phrases = True)
         
     solo_authors.delete_clones_of_citations(two_authors)
     
-    narrower_citations = PapersCited.CitationType(solo_authors.citations + 
+    narrower_citations = ca.CitationType(solo_authors.citations + 
                                       two_authors.citations +
                                       author_et_al.citations)
     
-    wider_citations = PapersCited.CitationType(three_authors.citations + 
+    wider_citations = ca.CitationType(three_authors.citations + 
                                    two_surnames.citations +
                                    two_surnames_et_al.citations)
 
     narrower_citations.cleanup()
     wider_citations.cleanup(allow_commas = False) # Default False prevents lots of duplication
     
-    expected_output = PapersCited.read_document(os.path.join(rootdir, "sample_text_full_correct_list_citations.txt"))
+    expected_output = ca.read_document(os.path.join(rootdir, "sample_text_full_correct_list_citations.txt"))
     expected_output = expected_output.split("\n")
     
     # When reading from .txt files, encoding/character issues arise.
@@ -40,23 +40,23 @@ def test_read_analyze_text(rootdir):
     assert narrower_citations.citations == expected_output
     
 def test_main_text_and_docx_footnotes_analyzed(rootdir):
-    text = PapersCited.read_document(os.path.join(rootdir, "document_footnotes.docx"))
+    text = ca.read_document(os.path.join(rootdir, "document_footnotes.docx"))
     
     # Get all types of citations
-    solo_authors = PapersCited.get_matches_solo_author(text, drop_excluded_phrases = True)
-    two_authors = PapersCited.get_matches_two_authors(text, drop_excluded_phrases = True)
-    three_authors = PapersCited.get_matches_three_authors(text, drop_excluded_phrases = True)
-    author_et_al = PapersCited.get_matches_author_et_al(text, drop_excluded_phrases = True)
-    two_surnames = PapersCited.get_matches_two_surnames(text, drop_excluded_phrases = True)
-    two_surnames_et_al = PapersCited.get_matches_two_surnames_et_al(text, drop_excluded_phrases = True)
+    solo_authors = ca.get_matches_solo_author(text, drop_excluded_phrases = True)
+    two_authors = ca.get_matches_two_authors(text, drop_excluded_phrases = True)
+    three_authors = ca.get_matches_three_authors(text, drop_excluded_phrases = True)
+    author_et_al = ca.get_matches_author_et_al(text, drop_excluded_phrases = True)
+    two_surnames = ca.get_matches_two_surnames(text, drop_excluded_phrases = True)
+    two_surnames_et_al = ca.get_matches_two_surnames_et_al(text, drop_excluded_phrases = True)
         
     solo_authors.delete_clones_of_citations(two_authors)
     
-    narrower_citations = PapersCited.CitationType(solo_authors.citations + 
+    narrower_citations = ca.CitationType(solo_authors.citations + 
                                       two_authors.citations +
                                       author_et_al.citations)
     
-    wider_citations = PapersCited.CitationType(three_authors.citations + 
+    wider_citations = ca.CitationType(three_authors.citations + 
                                    two_surnames.citations +
                                    two_surnames_et_al.citations)
 
@@ -68,23 +68,23 @@ def test_main_text_and_docx_footnotes_analyzed(rootdir):
     assert narrower_citations.citations == expected_output
 
 def test_docx_without_footnotes_analyzed(rootdir):
-    text = PapersCited.read_document(os.path.join(rootdir, "document_no_footnotes.docx"))
+    text = ca.read_document(os.path.join(rootdir, "document_no_footnotes.docx"))
     
     # Get all types of citations
-    solo_authors = PapersCited.get_matches_solo_author(text, drop_excluded_phrases = True)
-    two_authors = PapersCited.get_matches_two_authors(text, drop_excluded_phrases = True)
-    three_authors = PapersCited.get_matches_three_authors(text, drop_excluded_phrases = True)
-    author_et_al = PapersCited.get_matches_author_et_al(text, drop_excluded_phrases = True)
-    two_surnames = PapersCited.get_matches_two_surnames(text, drop_excluded_phrases = True)
-    two_surnames_et_al = PapersCited.get_matches_two_surnames_et_al(text, drop_excluded_phrases = True)
+    solo_authors = ca.get_matches_solo_author(text, drop_excluded_phrases = True)
+    two_authors = ca.get_matches_two_authors(text, drop_excluded_phrases = True)
+    three_authors = ca.get_matches_three_authors(text, drop_excluded_phrases = True)
+    author_et_al = ca.get_matches_author_et_al(text, drop_excluded_phrases = True)
+    two_surnames = ca.get_matches_two_surnames(text, drop_excluded_phrases = True)
+    two_surnames_et_al = ca.get_matches_two_surnames_et_al(text, drop_excluded_phrases = True)
         
     solo_authors.delete_clones_of_citations(two_authors)
     
-    narrower_citations = PapersCited.CitationType(solo_authors.citations + 
+    narrower_citations = ca.CitationType(solo_authors.citations + 
                                       two_authors.citations +
                                       author_et_al.citations)
     
-    wider_citations = PapersCited.CitationType(three_authors.citations + 
+    wider_citations = ca.CitationType(three_authors.citations + 
                                    two_surnames.citations +
                                    two_surnames_et_al.citations)
